% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{.onLoad}
\alias{.onLoad}
\title{Initial function}
\usage{
.onLoad(libname = base::find.package("SMRinR"), pkgname = "SMRinR")
}
\arguments{
\item{libname}{Path to package}

\item{pkgname}{Name of package}
}
\description{
Initial function
}
