% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getoutcome.R
\name{format_SMR}
\alias{format_SMR}
\title{Title Standardizing outcome format}
\usage{
format_SMR(
  file,
  sep = "\\t",
  SNP,
  A1,
  A2,
  freq,
  b,
  se,
  p,
  n = NULL,
  output = "mygwas"
)
}
\arguments{
\item{file}{The path or dataframe of outcome.}

\item{sep}{The delimiter of the input data.}

\item{SNP}{The column name of SNP.}

\item{A1}{The column name of alt.}

\item{A2}{The column name of ref.}

\item{freq}{The column name of af_alt.}

\item{b}{The column name of beta.}

\item{se}{The column name of sebeta.}

\item{p}{The column name of pval.}

\item{n}{The column name of ncase+ncontrol.}

\item{output}{The name of output file.}
}
\value{
A formatted file of outcome in current path.
}
\description{
Title Standardizing outcome format
}
