% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genelist.R
\name{GeneRowNum}
\alias{GeneRowNum}
\title{The functon of get gene list}
\usage{
GeneRowNum(GENELIST)
}
\arguments{
\item{GENELIST}{The list of gene}
}
\value{
the list of genes
}
\description{
The functon of get gene list
}
