% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_qlt.R
\name{get_gene_qlt}
\alias{get_gene_qlt}
\title{The function of making genes' qlt file}
\usage{
get_gene_qlt(qlt_path, genes, query = 5e-08, out_file)
}
\arguments{
\item{qlt_path}{qlt data's path}

\item{genes}{Gene list file, usually a gene symbol ID file or gene Ensembl id file}

\item{query}{Gene-associated SNP threshold, usually 5e-8}

\item{out_file}{output file path}
}
\description{
The function of making genes' qlt file
}
\details{
The demo code detailed in the test file
}
