% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getplotfile.R
\name{getplot}
\alias{getplot}
\title{The function for get plot file}
\usage{
getplot(ref = "EUR", gwas_path, qlt_path, probe, hgfile = "glist-hg19")
}
\arguments{
\item{ref}{the parameter of ref only can be choosed one of the following: EUR, SAS, AFR, AMR, and EAS}

\item{gwas_path}{outcome data' path}

\item{qlt_path}{qlt data's path}

\item{probe}{The probe which you want to plot}

\item{hgfile}{specifies a gene range list(glist-hg18,glist-hg19 or glist-hg38).}
}
\value{
plot data
}
\description{
The function for get plot file
}
