% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{initialize}
\alias{initialize}
\title{Initial function}
\usage{
initialize(libname, pkgname)
}
\arguments{
\item{libname}{Path to package}

\item{pkgname}{Name of package}
}
\description{
Initial function
}
