% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_plot.R
\name{plot_probe}
\alias{plot_probe}
\title{Plot a figure for probe}
\usage{
plot_probe(probeinfobuf, k, colplot, x.min, x.max, y.min, y.max, pchbuf, heidi)
}
\arguments{
\item{probeinfobuf}{probeinfobuf}

\item{k}{k}

\item{colplot}{colplot}

\item{x.min}{x.min}

\item{x.max}{x.max}

\item{y.min}{y.min}

\item{y.max}{y.max}

\item{pchbuf}{pchbuf}

\item{heidi}{heidi}
}
\description{
Plot a figure for probe
}
