% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smr.R
\name{runSMR}
\alias{runSMR}
\title{The function of running SMR in R}
\usage{
runSMR(ref = "EUR", gwas_path, qlt_path, maf = 0.01, out_file, thread = 4)
}
\arguments{
\item{ref}{the parameter of ref only can be choosed one of the following: EUR, SAS, AFR, AMR, and EAS}

\item{gwas_path}{outcome data' path}

\item{qlt_path}{qlt data's path}

\item{maf}{Numeric variable MAF (minor allele frequency)}

\item{out_file}{path of output file}

\item{thread}{the thread of work}
}
\description{
The function of running SMR in R
}
\details{
The demo code detailed in the test file
}
