% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smr_custom.R
\name{runSMR_custom}
\alias{runSMR_custom}
\title{The function of running SMR in R cutomically}
\usage{
runSMR_custom(command)
}
\arguments{
\item{ref}{The command of SMR software}
}
\description{
The function of running SMR in R cutomically
}
\details{
The demo code detailed in the test file
}
