% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{SMR_initialize}
\alias{SMR_initialize}
\title{Initialize the R package of "SMRinR"}
\usage{
SMR_initialize(file)
}
\arguments{
\item{file}{The path to the file of the R package initialization file}
}
\description{
Initialize the R package of "SMRinR"
}
\examples{
\donttest{
SMR_initialize('file path to initialization file')}
}
