% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMREffectPlot.R
\name{SMREffectPlot}
\alias{SMREffectPlot}
\title{The script of plotting SMREffectPlot}
\usage{
SMREffectPlot(
  data = SMRData,
  trait_name = "",
  cisWindow = 2000,
  transWindow = 5000,
  pointsize = 20
)
}
\arguments{
\item{data}{outcome data of SMR}

\item{trait_name}{name of outcome}

\item{cisWindow}{cisWindow}

\item{transWindow}{transWindow}

\item{pointsize}{pointsize}
}
\description{
The script of plotting SMREffectPlot
}
\examples{
# see "https://yanglab.westlake.edu.cn/software/smr/#SMRlocusplot19"
}
