% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRLocusPlot.R
\name{SMRLocusPlot}
\alias{SMRLocusPlot}
\title{The script of plotting SMR Locus Plot}
\usage{
SMRLocusPlot(
  data = SMRData,
  probeNEARBY = NULL,
  smr_thresh = NULL,
  smr_thresh_plot = NULL,
  heidi_thresh = NULL,
  plotWindow = NULL,
  pointsize = 20,
  max_anno_probe = 16,
  anno_selfdef = TRUE
)
}
\arguments{
\item{data}{outcome data of SMR}

\item{probeNEARBY}{probe}

\item{smr_thresh}{smr_thresh}

\item{smr_thresh_plot}{smr_thresh_plot}

\item{heidi_thresh}{thresh of heidi}

\item{plotWindow}{plotWindow}

\item{pointsize}{pointsize}

\item{max_anno_probe}{max_anno_probe}

\item{anno_selfdef}{anno_selfdef}
}
\description{
The script of plotting SMR Locus Plot
}
\examples{
# see "https://yanglab.westlake.edu.cn/software/smr/#SMRlocusplot19"

}
